kernel: add an error module
Most modules were getting crowded. Add a separate module for hosting
the error types.
