+++ src/kernel/instruction.rs
@@ -8,18 +8,10 @@ use std::{
 
 use log::debug;
 use serde::{Serialize, Serializer};
-use thiserror::Error;
 
 use crate::kernel::{REGVAL_SIZE_MASK, RegValSigned, instr_code_print::PrettyBincode};
 
-use super::{Memory, MemoryError, RegId, RegVal, Registers};
-
-/// Error returned by [`Instruction::execute`].
-#[derive(Clone, Copy, PartialEq, Eq, Error, Debug, Serialize)]
-pub enum InstructionError {
-    #[error("memory error: {0}")]
-    MemoryError(#[source] MemoryError),
-}
+use super::{InstructionError, Memory, RegId, RegVal, Registers};
 
 type Result<T> = std::result::Result<T, InstructionError>;
