+++ src/kernel/memory.rs
@@ -1,26 +1,8 @@
 //! Memory subsystem of the kernel.
 
 use log::{debug, info, trace};
-use serde::Serialize;
-use thiserror::Error;
 
-use crate::kernel::{InstrVal, RegVal};
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq, Error, Serialize)]
-pub enum MemoryError {
-    #[error("Address {address:#x} is not mapped")]
-    AddressOutOfRange { address: RegVal },
-    #[error("Address {address:#x}: segment doesn't allow reads")]
-    AddressNotReadable { address: RegVal },
-    #[error("Address {address:#x}: segment doesn't allow writes")]
-    AddressNotWritable { address: RegVal },
-    #[error("Address {address:#x}: segment doesn't allow execution")]
-    AddressNotExecutable { address: RegVal },
-    #[error("Address {address:#x} is not {expected_alignment}-aligned")]
-    MisalignedAccess { address: RegVal, expected_alignment: usize },
-    #[error("Segment {off:#x}:{len:#x} overlaps existing: {found_off:#x}:{found_len:#x}")]
-    SegmentOverlap { found_off: RegVal, found_len: RegVal, off: RegVal, len: RegVal },
-}
+use crate::kernel::{InstrVal, MemoryError, RegVal};
 
 type Result<T> = std::result::Result<T, MemoryError>;
